bpf: move LB-related structs to lb.h
Untangle common.h some more.
